WeatherInfoPresenter.kt deleted and functions moved to ViewModel
